requirements.txt
google-generativeai
chromadb
python-dotenv
tqdm
watchdog 
streamlit
sentence-transformers